Implemented AcceptShipment
Implemented the accept shipment process. It is not perfect, but it gets the job done. For now the user must know the string id of any product and the supplier id of any supplier to be able to enter a product. Maybe someday this will be changed so they can pick from a list